backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.1
requests==2.33.1
email-validator==2.1.0
# Legacy toolboxweb (EnrollRequestManager, AFC_SIMCard_Serach, Selenium crawlers)
numpy==2.2.4
pandas==2.2.3
werkzeug==3.1.3
openpyxl==3.1.5
selenium==4.27.1
websocket-client==1.9.0
jsonschema==4.23.0
psycopg2-binary==2.9.10
